+++ example/lib/inputs/go_text_input.dart
@@ -1,11 +1,15 @@
-
 import 'package:flutter/material.dart';
 import 'package:go_form/go_form.dart';
 import 'package:go_form_example/inputs/root_input.dart';
 
-class GoTextInput extends FormFieldModelBase<String>{
+class GoTextInput extends FormFieldModelBase<String> {
   final String label;
-  GoTextInput( {required super.name, super.validator,required this.label,});
+
+  GoTextInput({
+    required super.name,
+    super.validator,
+    required this.label,
+  });
 
   @override
   Widget build(BuildContext context, FieldController controller) {
@@ -15,7 +19,6 @@ class GoTextInput extends FormFieldModelBase<String>{
       validator: validator,
       errorText: controller.error,
       labelText: label,
-
     );
   }
 }