+++ example/lib/pages/home_page.dart
@@ -5,6 +5,7 @@ import 'package:go_form_example/pages/validate_listener_page.dart';
 import '../forms/forms.dart';
 import 'big_list_page.dart';
 import 'dynamic_actions_page.dart';
+import 'focus_listener_page.dart';
 import 'form_page.dart';
 import 'init_values_page.dart';
 
@@ -115,6 +116,18 @@ class _HomePageState extends State<HomePage> {
               );
             },
             child: Text('Validate Page'),
+
+          ),
+          ElevatedButton(
+            onPressed: () {
+              Navigator.push(
+                context,
+                MaterialPageRoute(
+                  builder: (context) => FocusListenerPage(),
+                ),
+              );
+            },
+            child: Text('Focus listener'),
 
           ),
         ],