+++ lib/src/domain/from_field_data.dart
@@ -3,7 +3,6 @@ import 'package:flutter/material.dart';
 class FormFieldData<T> {
   final ValueNotifier<T?> value;
   final String? error;
-  final FocusNode focusNode;
   final String? Function(T? value)? validator;
   final Key? key;
 
@@ -12,8 +11,7 @@ class FormFieldData<T> {
     this.validator,
     this.error,
     this.key,
-  })  : value = ValueNotifier<T?>(initialValue),
-        focusNode = FocusNode();
+  })  : value = ValueNotifier<T?>(initialValue);
 
   @override
   bool operator ==(Object other) {
@@ -38,7 +36,6 @@ class FormFieldData<T> {
   }
 
   void dispose() {
-    focusNode.dispose();
     value.dispose();
   }